refactor: split out inline cgo to it's own files
